# openapi.yaml
openapi: 3.1.0
info:
  title: basalt-server
  description: Basalt server competition runtime
  license:
    name: ''
  version: 0.1.0
paths:
  /announcements:
    get:
      tags:
      - announcements
      operationId: get_all
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Announcement'
    post:
      tags:
      - announcements
      operationId: new
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAnnouncement'
        required: true
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
        '401':
          description: User may not create announcements
  /announcements/{id}:
    delete:
      tags:
      - announcements
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/AnnouncementId'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
        '401':
          description: User may not delete announcements
        '404':
          description: Announcement with provided id does not exists
...
  /clock:
    get:
      tags:
      - clock
      operationId: get_clock
      responses:
        '200':
          description: Information about the clock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClockStatusResponse'
    patch:
      tags:
      - clock
      operationId: patch_clock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClockRequest'
        required: true
      responses:
        '200':
          description: Game is paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClockStatusResponse'
        '401':
          description: Insufficient permissions
...
components:
  schemas:
    Announcement:
      type: object
      required:
      - id
      - sender
      - time
      - message
      properties:
        id:
          $ref: '#/components/schemas/AnnouncementId'
        message:
          type: string
        sender:
          $ref: '#/components/schemas/UserId'
        time:
          type: string
          format: date
    AnnouncementId:
      type: string
    ClockStatusResponse:
      type: object
      required:
      - is_paused
      - time_left_in_seconds
      properties:
        is_paused:
          type: boolean
        time_left_in_seconds:
          type: integer
          format: int64
          minimum: 0
...
    NewAnnouncement:
      type: object
      required:
      - message
      properties:
        message:
          type: string
...
    UpdateClockRequest:
      oneOf:
      - type: object
        required:
        - is_paused
        properties:
          is_paused:
            type: boolean
...
    UserId:
      type: string